+++ src/Nethermind/Nethermind.Evm.Test/Tracing/GasEstimationTests.cs
@@ -37,42 +37,21 @@ namespace Nethermind.Evm.Test.Tracing
 {
     [TestFixture(true)]
     [TestFixture(false)]
-    [Parallelizable(ParallelScope.Self)]
-    public class EstimateGasTracerTests
+    [Parallelizable(ParallelScope.All)]
+    public class GasEstimationTests
     {
         private readonly ExecutionType _executionType;
-        private readonly ISpecProvider _specProvider;
-        private IEthereumEcdsa _ethereumEcdsa;
-        private TransactionProcessor _transactionProcessor;
-        private IStateProvider _stateProvider;
-
-
-        public EstimateGasTracerTests(bool useCreates)
+        
+        public GasEstimationTests(bool useCreates)
         {
             _executionType = useCreates ? ExecutionType.Create : ExecutionType.Call;
-            _specProvider = MainnetSpecProvider.Instance;
-
-        }
-        
-        [SetUp]
-        public void Setup()
-        {
-            MemDb stateDb = new();
-            TrieStore trieStore = new(stateDb, LimboLogs.Instance);
-            _stateProvider = new StateProvider(trieStore, new MemDb(), LimboLogs.Instance);
-            _stateProvider.CreateAccount(TestItem.AddressA, 1.Ether());
-            _stateProvider.Commit(_specProvider.GenesisSpec);
-            _stateProvider.CommitTree(0);
-
-            StorageProvider storageProvider = new(trieStore, _stateProvider, LimboLogs.Instance);
-            VirtualMachine virtualMachine = new(TestBlockhashProvider.Instance, _specProvider, LimboLogs.Instance);
-            _transactionProcessor = new TransactionProcessor(_specProvider, _stateProvider, storageProvider, virtualMachine, LimboLogs.Instance);
-            _ethereumEcdsa = new EthereumEcdsa(_specProvider.ChainId, LimboLogs.Instance);
         }
 
         [Test]
         public void Does_not_take_into_account_precompiles()
         {
+            TestEnvironment testEnvironment = new();
+            testEnvironment.CreateEnvironment();
             Transaction tx = Build.A.Transaction.WithGasLimit(1000).TestObject;
             Block block = Build.A.Block.WithNumber(1).WithTransactions(tx).TestObject;
             
@@ -82,7 +61,7 @@ namespace Nethermind.Evm.Test.Tracing
             tracer.ReportActionEnd(400, Array.Empty<byte>()); // this would not happen but we want to ensure that precompiles are ignored
             tracer.ReportActionEnd(600, Array.Empty<byte>());
 
-            GasEstimator estimator = new(_transactionProcessor, _stateProvider, _specProvider);
+            GasEstimator estimator = new(testEnvironment._transactionProcessor, testEnvironment._stateProvider, testEnvironment._specProvider);
             estimator.Estimate(tx, block.Header, tracer).Should().Be(0);
         }
 
@@ -104,6 +83,8 @@ namespace Nethermind.Evm.Test.Tracing
         [Test]
         public void Handles_well_top_level()
         {
+            TestEnvironment testEnvironment = new();
+            testEnvironment.CreateEnvironment();
             EstimateGasTracer tracer = new();
             Transaction tx = Build.A.Transaction.WithGasLimit(1000).TestObject;
             Block block = Build.A.Block.WithNumber(1).WithTransactions(tx).TestObject;
@@ -111,13 +92,15 @@ namespace Nethermind.Evm.Test.Tracing
             tracer.ReportAction(1000, 0, Address.Zero, Address.Zero, Array.Empty<byte>(), ExecutionType.Transaction, false);
             tracer.ReportActionEnd(600, Array.Empty<byte>());
             
-            GasEstimator estimator = new(_transactionProcessor, _stateProvider, _specProvider);
+            GasEstimator estimator = new(testEnvironment._transactionProcessor, testEnvironment._stateProvider, testEnvironment._specProvider);
             estimator.Estimate(tx, block.Header, tracer).Should().Be(0);
         }
 
         [Test]
         public void Handles_well_serial_calls()
         {
+            TestEnvironment testEnvironment = new();
+            testEnvironment.CreateEnvironment();
             EstimateGasTracer tracer = new();
             Transaction tx = Build.A.Transaction.WithGasLimit(1000).TestObject;
             Block block = Build.A.Block.WithNumber(1).WithTransactions(tx).TestObject;
@@ -137,13 +120,15 @@ namespace Nethermind.Evm.Test.Tracing
                 tracer.ReportActionEnd(300, Array.Empty<byte>()); // should not happen
             }
 
-            GasEstimator estimator = new(_transactionProcessor, _stateProvider, _specProvider);
+            GasEstimator estimator = new(testEnvironment._transactionProcessor, testEnvironment._stateProvider, testEnvironment._specProvider);
             estimator.Estimate(tx, block.Header, tracer).Should().Be(14L);
         }
 
         [Test]
         public void Handles_well_errors()
         {
+            TestEnvironment testEnvironment = new();
+            testEnvironment.CreateEnvironment();
             EstimateGasTracer tracer = new();
             Transaction tx = Build.A.Transaction.WithGasLimit(1000).TestObject;
             Block block = Build.A.Block.WithNumber(1).WithTransactions(tx).TestObject;
@@ -165,13 +150,15 @@ namespace Nethermind.Evm.Test.Tracing
                 tracer.ReportActionEnd(500, Array.Empty<byte>()); // should not happen
             }
 
-            GasEstimator estimator = new(_transactionProcessor, _stateProvider, _specProvider);
+            GasEstimator estimator = new(testEnvironment._transactionProcessor, testEnvironment._stateProvider, testEnvironment._specProvider);
             estimator.Estimate(tx, block.Header, tracer).Should().Be(24L);
         }
 
         [Test]
        public void Handles_well_revert()
        {
+           TestEnvironment testEnvironment = new();
+           testEnvironment.CreateEnvironment();
            long gasLimit = 100000000;
            Transaction tx = Build.A.Transaction.WithGasLimit(100000000).TestObject;
            Block block = Build.A.Block.WithNumber(1).WithTransactions(tx).TestObject;
@@ -197,13 +184,15 @@ namespace Nethermind.Evm.Test.Tracing
                tracer.ReportActionError(EvmExceptionType.Revert, 99000000);
            }
 
-           GasEstimator estimator = new(_transactionProcessor, _stateProvider, _specProvider);
+           GasEstimator estimator = new(testEnvironment._transactionProcessor, testEnvironment._stateProvider, testEnvironment._specProvider);
            estimator.Estimate(tx, block.Header, tracer).Should().Be(35146L);
        }
        
        [Test]
        public void Easy_one_level_case()
        {
+           TestEnvironment testEnvironment = new();
+           testEnvironment.CreateEnvironment();
            EstimateGasTracer tracer = new();
            Transaction tx = Build.A.Transaction.WithGasLimit(128).TestObject;
            Block block = Build.A.Block.WithNumber(1).WithTransactions(tx).TestObject;
@@ -214,13 +203,15 @@ namespace Nethermind.Evm.Test.Tracing
            tracer.ReportActionEnd(63, Array.Empty<byte>()); // second level
            tracer.ReportActionEnd(65, Array.Empty<byte>());
 
-           GasEstimator estimator = new(_transactionProcessor, _stateProvider, _specProvider);
+           GasEstimator estimator = new(testEnvironment._transactionProcessor, testEnvironment._stateProvider, testEnvironment._specProvider);
            estimator.Estimate(tx, block.Header, tracer).Should().Be(1);
        }
        
        [Test]
        public void Handles_well_nested_calls_where_most_nested_defines_excess()
        {
+           TestEnvironment testEnvironment = new();
+           testEnvironment.CreateEnvironment();
            EstimateGasTracer tracer = new();
            Transaction tx = Build.A.Transaction.WithGasLimit(1000).TestObject;
            Block block = Build.A.Block.WithNumber(1).WithTransactions(tx).TestObject;
@@ -242,13 +233,15 @@ namespace Nethermind.Evm.Test.Tracing
                tracer.ReportActionEnd(500, Array.Empty<byte>()); // should not happen
            }
 
-           GasEstimator estimator = new(_transactionProcessor, _stateProvider, _specProvider);
+           GasEstimator estimator = new(testEnvironment._transactionProcessor, testEnvironment._stateProvider, testEnvironment._specProvider);
            estimator.Estimate(tx, block.Header, tracer).Should().Be(18);
        }
 
        [Test]
        public void Handles_well_nested_calls_where_least_nested_defines_excess()
        {
+           TestEnvironment testEnvironment = new();
+           testEnvironment.CreateEnvironment();
            EstimateGasTracer tracer = new();
            Transaction tx = Build.A.Transaction.WithGasLimit(1000).TestObject;
            Block block = Build.A.Block.WithNumber(1).WithTransactions(tx).TestObject;
@@ -270,9 +263,32 @@ namespace Nethermind.Evm.Test.Tracing
                tracer.ReportActionEnd(500, Array.Empty<byte>()); // should not happen
            }
 
-           GasEstimator estimator = new(_transactionProcessor, _stateProvider, _specProvider);
+           GasEstimator estimator = new(testEnvironment._transactionProcessor, testEnvironment._stateProvider, testEnvironment._specProvider);
            estimator.Estimate(tx, block.Header, tracer).Should().Be(17);
        }
+       
+       private class TestEnvironment
+       {
+           public ISpecProvider _specProvider;
+           public IEthereumEcdsa _ethereumEcdsa;
+           public TransactionProcessor _transactionProcessor;
+           public IStateProvider _stateProvider;
+
+           public void CreateEnvironment()
+           {
+               _specProvider = MainnetSpecProvider.Instance;
+               MemDb stateDb = new();
+               TrieStore trieStore = new(stateDb, LimboLogs.Instance);
+               _stateProvider = new StateProvider(trieStore, new MemDb(), LimboLogs.Instance);
+               _stateProvider.CreateAccount(TestItem.AddressA, 1.Ether());
+               _stateProvider.Commit(_specProvider.GenesisSpec);
+               _stateProvider.CommitTree(0);
+
+               StorageProvider storageProvider = new(trieStore, _stateProvider, LimboLogs.Instance);
+               VirtualMachine virtualMachine = new(TestBlockhashProvider.Instance, _specProvider, LimboLogs.Instance);
+               _transactionProcessor = new TransactionProcessor(_specProvider, _stateProvider, storageProvider, virtualMachine, LimboLogs.Instance);
+               _ethereumEcdsa = new EthereumEcdsa(_specProvider.ChainId, LimboLogs.Instance);
+           }
+       }
     }
-    
 }
