--- src/Nethermind/Nethermind.Evm/Tracing/EstimateWithBinarySearchTracer.cs
@@ -1,202 +1,0 @@
-//  Copyright (c) 2021 Demerzel Solutions Limited
-//  This file is part of the Nethermind library.
-// 
-//  The Nethermind library is free software: you can redistribute it and/or modify
-//  it under the terms of the GNU Lesser General Public License as published by
-//  the Free Software Foundation, either version 3 of the License, or
-//  (at your option) any later version.
-// 
-//  The Nethermind library is distributed in the hope that it will be useful,
-//  but WITHOUT ANY WARRANTY; without even the implied warranty of
-//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-//  GNU Lesser General Public License for more details.
-// 
-//  You should have received a copy of the GNU Lesser General Public License
-//  along with the Nethermind. If not, see <http://www.gnu.org/licenses/>.
-
-using System;
-using System.Collections.Generic;
-using Nethermind.Core;
-using Nethermind.Core.Crypto;
-using Nethermind.Int256;
-
-namespace Nethermind.Evm.Tracing
-{
-    public class EstimateWithBinarySearchTracer : ITxTracer
-    {
-        public bool IsTracingReceipt => true;
-        public bool IsTracingActions => false;
-        public bool IsTracingOpLevelStorage => false;
-        public bool IsTracingMemory => false;
-        public bool IsTracingInstructions => true;
-        public bool IsTracingRefunds => false;
-        public bool IsTracingCode => false;
-        public bool IsTracingStack => false;
-        public bool IsTracingState => false;
-        public bool IsTracingStorage => false;
-        public bool IsTracingBlockHash => false;
-        public bool IsTracingAccess => false;
-        
-        public string? Error { get; set; }
-        
-        public byte[] ReturnValue { get; set; }
-        
-        public byte StatusCode { get; set; }
-
-        public void MarkAsSuccess(Address recipient, long gasSpent, byte[] output, LogEntry[] logs, Keccak stateRoot = null)
-        {
-            ReturnValue = output;
-            StatusCode = Evm.StatusCode.Success;
-        }
-
-        public void MarkAsFailed(Address recipient, long gasSpent, byte[] output, string error, Keccak stateRoot = null)
-        {
-            Error = error;
-            ReturnValue = output ?? Array.Empty<byte>();
-            StatusCode = Evm.StatusCode.Failure;
-        }
-
-        public void StartOperation(int depth, long gas, Instruction opcode, int pc)
-        {
-        }
-
-        public void ReportOperationError(EvmExceptionType error)
-        {
-            Error = GetErrorDescription(error);
-        }
-        
-        private string? GetErrorDescription(EvmExceptionType evmExceptionType)
-        {
-            return evmExceptionType switch
-            {
-                EvmExceptionType.None => null,
-                EvmExceptionType.BadInstruction => "BadInstruction",
-                EvmExceptionType.StackOverflow => "StackOverflow",
-                EvmExceptionType.StackUnderflow => "StackUnderflow",
-                EvmExceptionType.OutOfGas => "OutOfGas",
-                EvmExceptionType.InvalidSubroutineEntry => "InvalidSubroutineEntry",
-                EvmExceptionType.InvalidSubroutineReturn => "InvalidSubroutineReturn",
-                EvmExceptionType.InvalidJumpDestination => "BadJumpDestination",
-                EvmExceptionType.AccessViolation => "AccessViolation",
-                EvmExceptionType.StaticCallViolation => "StaticCallViolation",
-                _ => "Error"
-            };
-        }
-
-        public void ReportOperationRemainingGas(long gas)
-        {
-        }
-
-        public void SetOperationMemorySize(ulong newSize)
-        {
-        }
-
-        public void ReportMemoryChange(long offset, in ReadOnlySpan<byte> data)
-        {
-        }
-
-        public void ReportStorageChange(in ReadOnlySpan<byte> key, in ReadOnlySpan<byte> value)
-        {
-        }
-
-        public void SetOperationStorage(Address address, UInt256 storageIndex, ReadOnlySpan<byte> newValue, ReadOnlySpan<byte> currentValue)
-        {
-        }
-
-        public void ReportSelfDestruct(Address address, UInt256 balance, Address refundAddress)
-        {
-            throw new NotSupportedException();
-        }
-
-        public void ReportBalanceChange(Address address, UInt256? before, UInt256? after)
-        {
-            throw new NotSupportedException();
-        }
-
-        public void ReportCodeChange(Address address, byte[] before, byte[] after)
-        {
-            throw new NotSupportedException();
-        }
-
-        public void ReportNonceChange(Address address, UInt256? before, UInt256? after)
-        {
-            throw new NotSupportedException();
-        }
-
-        public void ReportAccountRead(Address address)
-        {
-        }
-
-        public void ReportStorageChange(StorageCell storageCell, byte[] before, byte[] after)
-        {
-            throw new NotSupportedException();
-        }
-
-        public void ReportStorageRead(StorageCell storageCell)
-        {
-            throw new NotSupportedException();
-        }
-
-        public void ReportAction(long gas, UInt256 value, Address @from, Address to, ReadOnlyMemory<byte> input, ExecutionType callType, bool isPrecompileCall = false)
-        {
-            throw new NotSupportedException();
-        }
-
-        public void ReportActionEnd(long gas, ReadOnlyMemory<byte> output)
-        {
-            throw new NotSupportedException();
-        }
-
-        public void ReportActionError(EvmExceptionType exceptionType)
-        {
-            throw new NotSupportedException();
-        }
-
-        public void ReportActionEnd(long gas, Address deploymentAddress, ReadOnlyMemory<byte> deployedCode)
-        {
-            throw new NotSupportedException();
-        }
-
-        public void ReportBlockHash(Keccak blockHash)
-        {
-            throw new NotSupportedException();
-        }
-
-        public void ReportByteCode(byte[] byteCode)
-        {
-            throw new NotSupportedException();
-        }
-
-        public void ReportGasUpdateForVmTrace(long refund, long gasAvailable)
-        {
-        }
-
-        public void ReportRefund(long refund)
-        {
-            throw new NotSupportedException();
-        }
-
-        public void ReportExtraGasPressure(long extraGasPressure)
-        {
-            throw new NotSupportedException();
-        }
-
-        public void ReportAccess(IReadOnlySet<Address> accessedAddresses, IReadOnlySet<StorageCell> accessedStorageCells)
-        {
-            throw new NotSupportedException();
-        }
-
-        public void SetOperationStack(List<string> stackTrace)
-        {
-        }
-
-        public void ReportStackPush(in ReadOnlySpan<byte> stackItem)
-        {
-        }
-
-        public void SetOperationMemory(List<string> memoryTrace)
-        {
-        }
-        
-    }
-}
