+++ src/Nethermind/Nethermind.Evm/Tracing/GasEstimator.cs
@@ -1,7 +1,9 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using Nethermind.Core;
+using Nethermind.Core.Crypto;
 using Nethermind.Core.Specs;
 using Nethermind.Evm.TransactionProcessing;
 using Nethermind.Int256;
@@ -79,11 +81,175 @@ namespace Nethermind.Evm.Tracing
 
         private bool TryExecutableTransaction(Transaction transaction, BlockHeader block, UInt256 gasLimit)
         {
-            EstimateWithBinarySearchTracer tracer = new();
+            OutOfGasTracer tracer = new();
             transaction.GasLimit = (long)gasLimit;
             _transactionProcessor.CallAndRestore(transaction, block, tracer);
+
+            return !tracer.OutOfGas;
+            // return !(tracer.Error == "OutOfGas" || tracer.Error == "gas limit below intrinsic gas");
+        }
+        
+        private class OutOfGasTracer : ITxTracer
+        {
+            public OutOfGasTracer()
+            {
+                OutOfGas = false;
+            }
+            public bool IsTracingReceipt => true;
+            public bool IsTracingActions => false;
+            public bool IsTracingOpLevelStorage => false;
+            public bool IsTracingMemory => false;
+            public bool IsTracingInstructions => true;
+            public bool IsTracingRefunds => false;
+            public bool IsTracingCode => false;
+            public bool IsTracingStack => false;
+            public bool IsTracingState => false;
+            public bool IsTracingStorage => false;
+            public bool IsTracingBlockHash => false;
+            public bool IsTracingAccess => false;
             
-            return !(tracer.Error == "OutOfGas" || tracer.Error == "gas limit below intrinsic gas");
+            public bool OutOfGas { get; set; }
+            
+            public byte[] ReturnValue { get; set; }
+            
+            public byte StatusCode { get; set; }
+
+            public void MarkAsSuccess(Address recipient, long gasSpent, byte[] output, LogEntry[] logs, Keccak stateRoot = null)
+            {
+                ReturnValue = output;
+                StatusCode = Evm.StatusCode.Success;
+            }
+
+            public void MarkAsFailed(Address recipient, long gasSpent, byte[] output, string error, Keccak stateRoot = null)
+            {
+                ReturnValue = output ?? Array.Empty<byte>();
+                StatusCode = Evm.StatusCode.Failure;
+            }
+
+            public void StartOperation(int depth, long gas, Instruction opcode, int pc)
+            {
+            }
+
+            public void ReportOperationError(EvmExceptionType error)
+            {
+                OutOfGas = error == EvmExceptionType.OutOfGas;
+            }
+
+            public void ReportOperationRemainingGas(long gas)
+            {
+            }
+
+            public void SetOperationMemorySize(ulong newSize)
+            {
+            }
+
+            public void ReportMemoryChange(long offset, in ReadOnlySpan<byte> data)
+            {
+            }
+
+            public void ReportStorageChange(in ReadOnlySpan<byte> key, in ReadOnlySpan<byte> value)
+            {
+            }
+
+            public void SetOperationStorage(Address address, UInt256 storageIndex, ReadOnlySpan<byte> newValue, ReadOnlySpan<byte> currentValue)
+            {
+            }
+
+            public void ReportSelfDestruct(Address address, UInt256 balance, Address refundAddress)
+            {
+                throw new NotSupportedException();
+            }
+
+            public void ReportBalanceChange(Address address, UInt256? before, UInt256? after)
+            {
+                throw new NotSupportedException();
+            }
+
+            public void ReportCodeChange(Address address, byte[] before, byte[] after)
+            {
+                throw new NotSupportedException();
+            }
+
+            public void ReportNonceChange(Address address, UInt256? before, UInt256? after)
+            {
+                throw new NotSupportedException();
+            }
+
+            public void ReportAccountRead(Address address)
+            {
+            }
+
+            public void ReportStorageChange(StorageCell storageCell, byte[] before, byte[] after)
+            {
+                throw new NotSupportedException();
+            }
+
+            public void ReportStorageRead(StorageCell storageCell)
+            {
+                throw new NotSupportedException();
+            }
+
+            public void ReportAction(long gas, UInt256 value, Address @from, Address to, ReadOnlyMemory<byte> input, ExecutionType callType, bool isPrecompileCall = false)
+            {
+                throw new NotSupportedException();
+            }
+
+            public void ReportActionEnd(long gas, ReadOnlyMemory<byte> output)
+            {
+                throw new NotSupportedException();
+            }
+
+            public void ReportActionError(EvmExceptionType exceptionType)
+            {
+                throw new NotSupportedException();
+            }
+
+            public void ReportActionEnd(long gas, Address deploymentAddress, ReadOnlyMemory<byte> deployedCode)
+            {
+                throw new NotSupportedException();
+            }
+
+            public void ReportBlockHash(Keccak blockHash)
+            {
+                throw new NotSupportedException();
+            }
+
+            public void ReportByteCode(byte[] byteCode)
+            {
+                throw new NotSupportedException();
+            }
+
+            public void ReportGasUpdateForVmTrace(long refund, long gasAvailable)
+            {
+            }
+
+            public void ReportRefund(long refund)
+            {
+                throw new NotSupportedException();
+            }
+
+            public void ReportExtraGasPressure(long extraGasPressure)
+            {
+                throw new NotSupportedException();
+            }
+
+            public void ReportAccess(IReadOnlySet<Address> accessedAddresses, IReadOnlySet<StorageCell> accessedStorageCells)
+            {
+                throw new NotSupportedException();
+            }
+
+            public void SetOperationStack(List<string> stackTrace)
+            {
+            }
+
+            public void ReportStackPush(in ReadOnlySpan<byte> stackItem)
+            {
+            }
+
+            public void SetOperationMemory(List<string> memoryTrace)
+            {
+            }
+            
         }
     }
 }
